separate Induct class by dependency
